requirements.txt
Django==1.11.17
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
python-social-auth==0.3.6
social-auth-app-django
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.6.1 
mysql-python
ipython
djangorestframework
